requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
scikit-learn
pandas
joblib
pydantic
pytest
